requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-docx==1.1.0
PyPDF2==3.0.1
pdfplumber==0.10.3
openpyxl==3.1.2
python-dateutil==2.8.2
werkzeug==3.0.1
jinja2==3.1.2
